gpuav: Rename Utils to StateTracker
