+++ BUILD.gn
@@ -161,8 +161,8 @@ core_validation_sources = [
   "layers/vulkan/generated/sync_validation_types.cpp",
   "layers/vulkan/generated/sync_validation_types.h",
   "layers/gpu_shaders/gpu_shaders_constants.h",
-  "layers/gpu_validation/gpu_utils.cpp",
-  "layers/gpu_validation/gpu_utils.h",
+  "layers/gpu_validation/gpu_state_tracker.cpp",
+  "layers/gpu_validation/gpu_state_tracker.h",
   "layers/gpu_validation/gpu_error_message.cpp",
   "layers/gpu_validation/gpu_error_message.h",
   "layers/gpu_validation/gpu_validation.cpp",
