+++ layers/gpu_validation/debug_printf.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "gpu_validation/gpu_utils.h"
+#include "gpu_validation/gpu_state_tracker.h"
 #include "gpu_validation/gpu_error_message.h"
 class DebugPrintf;
